pazienteService, pazienteDTO

+++ src/main/java/it/capstone/arno/repository/AnagraficaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnagraficaRepository extends JpaRepository<Anagrafica, Long> {
+public interface AnagraficaRepository extends JpaRepository<Anagrafica, Integer> {
 }
