+++ src/main/java/it/capstone/arno/repository/CartellaClinicaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartellaClinicaRepository extends JpaRepository<CartellaClinica, Long> {
+public interface CartellaClinicaRepository extends JpaRepository<CartellaClinica, Integer> {
 }
